Revamped the multilocus genotype prob calculation
I put a running log calculation in there to make sure that we don't underflow.

I tested it on a data set with 500 loci that was causing underflow before and
it is no longer underflowing.  Also compared results with the old version on a
few data sets and am getting identical results (well, slight differences after
the 16th decimal point or so...but that is to be expected.