Change call to MixtureSim to avoid (null)
I hacked MixtureSim a while ago to allow the sketchy multi-mix option.  When I did that I changed the way it prints things out, but I needed to pass "" instead of NULL to the new args I had added (cuz otherwise I got a lot of "(null)"s in the output.  Seems to be fixed now.